GUI which adds entries from different excel sheets
The add_separate_sheets tool is a tkinter GUI which allows a user to add entries lying in a single excel file, but on separate sheets. The user is prompted to pick a file by browsing, and then inputs which cells they want added up. The program also allows a user to download a .csv file with the obtained result. Also uploaded a screenshot of the program in action.